# src/main/resources/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Marketplace API
  version: 1.0.0
  description: API for marketplace
servers:
  - url: /api


#используемые объекты
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - name
      properties:
        email:
          type: string
          format: email
          description: Email пользователя
        password:
          type: string
          minLength: 6
          maxLength: 255
          description: Пароль
        name:
          type: string
          minLength: 2
          maxLength: 100
          description: Имя пользователя
...
    RegisterResponse:
      type: object
      properties:
        message:
          type: string
          example: User registered successfully
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
            email:
              type: string
              format: email
            name:
              type: string
            role:
              type: string
              enum: [ ROLE_USER, ROLE_SELLER, ROLE_ADMIN ]
            created_at:
              type: string
              format: date-time
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: Bearer
        user:
          type: object
          nullable: true
          properties:
            id:
              type: string
              format: uuid
            email:
              type: string
            name:
              type: string
            role:
              type: string
    RefreshTokenRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string

    # СХЕМЫ ЗАКАЗА
...
    ErrorResponse:
      type: object
      properties:
        error_code:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true


#ПУТИ
paths:
...
  /auth/register:
    post:
      tags:
        - Auth
      summary: Регистрация нового пользователя
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Пользователь зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Ошибка валидации VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: [ ]
...
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Обновление access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Токен обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Невалидный refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: [ ]

  #ПУТИ ЗАКАЗОВ
...